+++ common/src/main/java/com/flight_booking/common/infrastructure/security/AuthenticationFilter.java
@@ -1,40 +1,41 @@
-package com.flight_booking.user_service.infrastructure.security.authentication;
+package com.flight_booking.common.infrastructure.security;
 
-import com.flight_booking.user_service.presentation.global.exception.ErrorCode;
+import com.flight_booking.common.domain.model.Role;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-@Component
 @Slf4j(topic = "사용자 인증 필터")
-@RequiredArgsConstructor
 public class AuthenticationFilter extends OncePerRequestFilter {
-
-  private final CustomUserDetailsService customUserDetailsService;
 
   @Override
   protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
       FilterChain filterChain) throws ServletException, IOException {
 
-    log.info("--------user-service------------");
+    // 헤더에서 사용자 정보 추출
     String email = request.getHeader("X-USER-EMAIL");
-    String role = request.getHeader("X-USER-ROLE");
-    log.info("header: {}, {}", email, role);
+    String header = request.getHeader("X-USER-ROLE");
 
-    if (email != null && role != null) {
+    log.info("header: {}, {}", email, header);
+
+    if (email != null && header != null) {
       try {
+        // "ROLE_" 접두어 제거
+        String roleName = header.replace("ROLE_", "");
+
+        // Role enum에 맞게 변환
+        Role role = Role.valueOf(roleName);
+
         // 이메일로 사용자 정보 로드
-        UserDetails userDetails = customUserDetailsService.loadUserByUsername(email);
+        UserDetails userDetails = new CustomUserDetails(email, role);
 
         // 인증 객체 생성 (권한 포함)
         Authentication authentication = new UsernamePasswordAuthenticationToken(
@@ -48,9 +49,8 @@ public class AuthenticationFilter extends OncePerRequestFilter {
         // 사용자 정보 로드 실패 시, 에러 처리
         log.error("사용자 인증 실패: {}", e.getMessage());
 
-        ErrorCode errorCode = ErrorCode.USER_AUTHENTICATION_FAILED;
-        response.setStatus(errorCode.getHttpStatus().value());
-        response.getWriter().write(errorCode.getMessage());
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        response.getWriter().write("사용자 인증 실패");
         return;
       }
     }
